agents/epoch_breaker_expert_agent/requirements.txt
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
azure-ai-agentserver-agentframework==1.0.0b10
azure-ai-agentserver-core==1.0.0b10
azure-identity==1.17.1
python-dotenv==1.0.1
